Evaluate lambda functions
This is a first run at evaluating function constructs and is in some ways still
a work in progress.  In particular the implimentation in this commit does not
allow for any changes to the value of variables in the outer scope once a
function is defined -- though it does have access to these variables.  There is
also no way to effect the outer enviroment from within a function.  This makes
functions somewhat pure.  They can perform IO through built-in functions and can
read from their surrounding envrionment but cannot make changes to it.

Because of current parser limitations, only fixed arity functions can be
defined.

TODO
 * Block parsing
 * optional parameters and catch all parameters need to be parsed and evaluated
   correctly.

Notice: This commit is the first to contain the Author's Log.