Block Parsing and Evaluation
Blocks can now be parsed and evaluated resulting in their last value.

In order to parse multi-line expressions it was nesecary to consider
the line ending marker TEnd which is added in the scanLine function
for lines without blocks.  When parsing tokens that represent multiple
lines it is necessary to look for the TEnd or TBlock.  Unfortunatly,
this caused a bit of confusion with assignments with statements as the
RHS.  Parsing the RHS consumed the block but the original expression
saw neither the block nor a TEnd.

To solve this problem, a Bool userState was added to the parser
(technically making it no longer context free).  When this is set a
block has been consumed and there is no need to look for a TEnd.

With block parsing in place we have switched over to a multi-line
parser as the default for parseString.

Parser error messages were cleaned up.

REPL:
added a parser mode that will parse, but not evaluate the expressions.

TODO:
  * add a tokens mode to REPL
  * start on classes and objects
  * parse non-parened arguments and paramenters
  * explore continuation passing style with continuation in context
    for functions to facalitate `return` statements.