Moved unrealted definations out out Object.hs
The Object.hs file (and the AST.hs file before that) were full of type
and data definations that didn't really belong in them because it
prevented recursive mutual dependancies.  This commit moves those
dependancies into the correct files and uses GHC's {-# SOURCE #-}
programa to break the cycles.

Eval and Object need to know about each other because of the
defination of Fn.

A Longer cycle is created when Eval needs to drive Context, Context
need Object and Object needs Eval.

In breaking these cycles it became obvious that Var (a variable name
with scope) needed to be its own thing.  It therefore spawned a
seperate module with no external dependancies.

The changes in all other files are strictly updateing import
statements to reflect the new arrangements.

TODO:
  * context need to handle scoped variables corectly rather then
    ignoring them.
  * Need an initial Object class and main instance for Main.hs
  * Work out how presedence with be stored.
  * Evaluate class
  * `new` and `spawn` to make new instances.