Initial Eval of Class Expressions [WIP]
Classes can now be instntiated, at least to the level of making a
value for them.  The need now is to create the initialObject graph and
add the functions to travirse that graph.

This commit required a lot of fussing to get break cyclical
dependancies.  Note the large number of `*.hs-boot` files created.

Object.Spawn is also added and it the first piece of concurrency to
make it into Sapphire.