Object Graph Search
We can now search the object graph per the rules described in the
authors log dated December 9, 2013 at 9:52am

It has become easier to merge PID's into the Object data type.  Nearly
everywhere an Object was called for there was a like about Either Pid
Object. The solution was to merge the two types.

TODO:
  * Create Object and main and use them to complete the Main.hs
    initial context.
  * Create the default objects
  * impliment instance variable in lexer, parser and eval (@foo)