Lex, Parse, Eval `self`; towards class names [WIP]
We now can deal with `self` all the way throug the pipe-line.

In attempting to alow classes to be reference by their proper name, we
have run into a problem with our current message queue.  AS expected loops
in the line of passed messages create a dead-lock.  We must address the
dead-lock before moving forward.

[WIP]