*Bug*: Deadlocks occuring
See the Author's log for full details.

I have marked the lines that are causing the deadlocks. This commit is to
be the root for some expirimental branches one using MVars one converting
to STM. I may also start a branch where I kick the can down the road and
complete the object graph as that may be necessary to flesh out the
requirements of Pid message queues.