STM Conversion complete to typecheck
All occururances of Chan and MVar have been replaced with their STM
equivelent.  All files typecheck and Main compiles to an exicutable which
we assume still suffers from the same deadlock as the non-STM version.

NOTE: Object graph searches are now in the STM monad not the IO monad.
This makes the composible with the STM combinators, but requires the
initial search to user `atomically` to lift the STM inot the IO monad.