Continuations [WIP]
The logic for sending messages and recieving responses has been refactored
out of Object.hs and into Continuation.hs.

Spawn, Graph and Eval have not been updated to reflect the change, the
only requirement for this commit was that Continuation, which has not
dependacies, compiles, and the Object compiles to a .o file (dependancies
still are broken.

Continuation introduces a new Monad, `ContM` which is the continuation
state wrapping the STM monad.