STM Working to Compile
The new STM continuation system that was the goal of the STM branch into
which this commit is made now compiles along with all the major pieces of
the interperter.

The next set is to create the runtime environment and get back to a
minimually working interperter.  The Boot.hs file has jey yo be touched
and lines 35 - 41 will need major reworking to comply with the new
continuation and context regime.

It is expected that making these changes may reveal further work needed in
the STM COntinuation layout.  but given that they are currently
well-formed per GHC this is a good place to make a commit so that if
everything gets really messed up we can come back to this stable point.

TODO:
  * There are two undefined functions that need to be worked out.  One is
    the getMethod function in Object.Spawn.  There were major changes to
    the way the Object graph search in implimented.  There is still some
    refactoring work in that area of the code (it is still unclear what
    belongs in Context and what in Object Graph) A commit was needed
    before attempting this refactor.
  * In Context the call to dispatch intentionally has been left with
    `undefined` as the Responder. It is unclear whether the idea of simply
    setting up a responder on Self is sufficient if Self is a local
    sub-object.  This also dealys the need to fix a cyclic dependance
    between Context and Object.Spawn
  * Fix Boot and likely other runtime pieces for the new STM Continuations
  * Eval of call and apply using sendM and dispatchM as appropriate
  * Proper tail calls for Blocks. (!)

Notes:
  Once STM is working, it would be best to merge back with the main
  branch before adding to the runtime system further. The current
  (non-working) stuff in Boot is enough to show that STM can avoid looping
  call graphs.