Added Cabal file
Added a cabal file to the project with version set to 0.1.0.0.  I have
chosen to licence this project under the GNU GPLv3.0 and should add a
notice to the top of all my files, but for now we'll fly under the radar.

Further I need to clairify that the /doc directory is not covered by the
GPL.

Also, it is my intent that like gcc and other compilers, the source files
and output (someday) of sapphire is not covered by my copyright only the
program itself.  (Writing a compiler for a concurrent language is easy
compaired to copyright law. bleck!)

There is also a minor annoyance that Tokens.x is not compiled with the -g
flag.  At present GCH is the only complier capible of compiling this
source due to language extensions.  The -g option in Alex vastly improves
performance of both the compiler and the generated code.  A wise step
would be to hand compile using `alex -g Tokens.x` before building until I
can resolve the issue.