Rewrite of REPL, load file
The REPL in main.hs has been rewritten to allow for more sophisticated
parsing of system commands.  In particular we can now load files to be
scanned, parsed, or evaluated.  There is also a help screen from the
system command prompt

In working on this, I have uncovered a bug in the parser.  For reasons I
have not yet gotten to the bottom of, a blank line causes a parsing error.