BUG FIX #17:  Blank line borks parser
This bug was not in the parser but in the line parser.  Blank lines were
not accounted for in a few places and the resulted in errors.

In particular the edge chase of a blank line starting a block was un
handled.  The assumption was that a blank line was just part of the
surrounding block.  But if the next line was indented, the line parser
would try to add the block to a BlankLine which does not have a `block`
field.

A simular problem was encountered in Tokens wher an attempt to scan a
blank line resulted in any blank line borking the whole program.

This got blamed on the parser thanks to Haskell's lazy evaluation.  Until
the parser tried to analyze the first token the undefined condition was
not evaluated.

testfile3 was modifies to show these changes result in correct behavior.
run the executable, go to the system prompt and type:

    load testfile3

to see the result.

Other changes: Documentation for Boot.hs