Graph Search Rewrite Complete
The object graph is formed from the `klass` and `super` fields of an object.
Object/Graph.hs provides functions to earch through the forrest of bojects for
ivars and cvars.  Prior to this rewrite, these functions had been scattered
about in Object.Spawn, Object.Graph and Context and the naming was both
confusing and inconsistant.

This commit caps a series of changes starting from 1a3ed4172 that rationalizes
the naming scheme and moves all the relevent functions into Object.Graph.

NOTE: one gottcha is that while lookupIVars never traverses the graph,
lookupCVars always does.  Further some of the logic for this is still bleading
into Spawn.hs because of the need to proper tail calls.  A future refactoring
should look to address this issue.  (perhaps at the same time modules are
implimented.)

As part of this process addational files have recieved haddock documentation and
copyright notices.  This is an on-going process.