Framework for Adding Built in Class Functions
Classes with built in representations (i.e. String, Array, Hash, etc) need a
way to connect their internal functions into their Sapphire Class, and indeed
need to create and install their classes.  Doing so requires an active context;
that is it needs to occure after `boot` is called.

In the interest of keeping most initialization out of the main function, we
have chosen to add a single function, `__initialize`, to the installed
functions in Object.  This function is called before anything else in base.sap
and will initialize all the built in classes using their provided
initialization functions.

These functions have access to the a valid, though very limited, Context within
the EvalM monad.  They will need to create a new class and add it to Object
using low level functions.  This may become a place for refactoring.

Within this commit, Array is so initialized with a single function, `length`.
Addational function need to be added.