Foldable instance for Array
Arrays are now instance of `Data.Foldable`.  To accomplish this Arrays
needed to become true containers so a datatype Array` was created to hold
generic values and Array was redefined as an allias for an Array' holding
Values.

This should allow us to impliment the arrays built in traversal function
easily.