Array Internals Re-write
After much expirimentation and frustration implimenting arrays that
exibited O(1) behavior for indexing, I have come to the conclusion that
the performance given by the `Data.Sequence.Seq` datatype is sufficient
for Sapphire.

This has greatly simplified the implimentation of Array as it is now
merely a `Seq Value` with all the accompanying function from
`Data.Sequence` re-exported.

Basic indexing, itteration through inject, and length are now working.  At
this point it seems quite worthwhile to return to the main tree and
impliment function blocks which are quite handy when dealing with arrays.

Further Notes on Data.Sequence
==============================

Data.Sequence is an implimentation of 2-3 finger trees annotated with
sizes as described in Ralf Hinze and Ross Paterson, "Finger trees: a
simple general-purpose data structure", Journal of Functional Programming
16:2 (2006) pp 197-217. (Section 4.2)
http://www.soi.city.ac.uk/~ross/papers/FingerTree.html

Index operations are O(log(min(i,n-i))) that is, the operation takes
logrithmic with respect to the distance the element is from the edge of
the sequence.