Parse `do` blocks
Updated the parser to accept a "do block" in the appropriate places.
Functionally, this adds anonymous function as the final argument to the
function.

WARNING: At present the interperter does not close over blocks. This
common behavior can cause problems if not implimented carefully. However,
this is now a known bug.