Addational builtin / derived functions
More built in and derived functions for Array and Hash.

Attempting to write some of these functions has revieled that while Array
and Hash were needed in order to describe some needed features, that the
language it's self is not complete or stable enough to support any kind of
extensive work in the area of Class implimentation.

A new Milestone has been created on Github with the goal of firming up the
base language enough that a more complete build out of derived methods
woudl be worth the effort.