Fixed Call Infinite Loop
This closes Issue #37

Missing functions became `nil.call` as "call" was not a method on nil the
cycle was able to repete.  Now `nil` spesifically fails as a reciever of
method.  Once `method_missing` is complete we can throw an error instead.

Also in this commit, some debugging facilities have been added that allow
introspection of objects able to recieve a call.  It is also possible to
view the local scope variables with `debug` from the system prompt.