Refactor of Builtin Function
Primitive types have functions that are written in pure haskell code.  The
code to construct the classes for these types are now consilidated into a
`Builtins` directory. This also allowed for a much needed clean-up of the
Boot.hs file.

Other local cleanup has been done in the course of this work with some
cruft removed.