Added load Command to Object
The method `load` is now a built in command of Object it reads the file
locates in its first argument and evaluates it in the current context.

There is also some clean up to the boot process greatly simplifying the
`main` function.  The date and time of the most recent compile are printed
when the interpreter is loaded.  Idealy this would be replaced with a
version number and copyright notice once Sapphire is released.