File Path now set in __FILE__
The local variable __FILE__ is now consistantly set to the path of the
currently executing file.  At boot it is set to the path of the
interpreter.

There is some question in my mind about what happends at the end of an
evalT invocation.  I believe that a tail call will wipe out the current
local context in favor of the invoked continuation's so the the __FILE__
variable will reset.  I have nto been able to test this logic and if a
strange bug with __FILE__ showing the location of a previously invoked
method turns up this is the cause.

We cannot simply reset the path in the above case because it would break
tail calls.