Rationalize Layout of Standard Library
With a working `load` function, it is now possible to begin work in ernist
on creating a standard library.  To this end, we have moved sapphire code
into lib and started a subdirectory set.

Before true work can begin, there must be enough functionality to have a
working `require` function.  In begining to write this function it became
clear that there were still a few missing pieces.  I made this commit so
that I could work on those pieces.