Method Visibility Added
Closes issue #53.

The consept of Public/Private visibility has been added.  Functions (and
other objects) in `ivars` are now considered private and cannot be called
with an explisit reciever.

Because we used ivars to hold some built in class methods and some very
primitive bootstrapping functions (e.g. `setClass` used to tie the
recursivaly defined knot of `Object` and `Class`) it was necessary to
rewrite a large amount of code.

To make publically visiable class methods, it is now necessary to put
those methods in an anonomous module just above the class.  We have chosen
to give such modules the proper name `*`.  This mechinism will aslo be
useful in addressing issue #51.

TODO:
Refactor Eval so that Call defers to Apply in all cases not just remote
objects.