Bug Fix: Instances not updating with their class
When the class of an instance was altered after that instance was created,
those alterations were not visible from the old instance but would be
visible to all new instancese.

This turned out to be a bug in `Class::new` which set `klass` of the new
instance equal to a copy of the classes structure rather then the PID of
the class.