REFACTOR: Call defers to Apply
Evaluating a `Call` node now always delegates action to `Apply`

After finding the reciever at run time, the reciever is placed into the
current context and an equivelent Apply node is created and evaluated.

This change keeps the logic for finding and applying function in one place
in the code.