WIP: converted Object{process} to TMVar
Altered the type of the `process` field of `Object` to be `TMVar Process`.
This has the effect of giving every object a unique identity (almost) for
free.  Many files have been changed to reflect this alteration.

Because we are using a TMVar it is no longer possible to construct an
object outside the IO monad. However because of the known safe nature of
`newEmptyTMVarIO` (see documentation) we have wish a certian shame
relented and begun using `unsafeperformIO`.  We feel so impure.