WIP: Function `with` removes scope changes
In Context, the `with` function is used to chnge which object is "self" it
can now also be used to revery changes to scope and global fields.  This
means that with appropriate changes to `eval` (planned) it should be
possible to set up the scope and even global values run a method, and
return to the original state.

The trick is propigating these changes appropriatly accross process
boundries.  I believe this will require changes in spawn and the message
structure in Object to accomodate.