WIP: `scope` and `global` in Responder
When a new process is spawned, `scope` and `global` are now included in
the responderObject function.

No function yet tries to read or make use of these values.

This WIP compiles and runs because noting yet uses these fileds, but
whether they have been set correctly still need to be checked.