Module and Class Scope
Classes and modules can now be defined within other modules allowing for
namespases.  The scope operator, `::`, now works as expected.

This change has required a extension of Context to hold a `scope` (where
new classes are installed and where undecorated constants are sought) and
`global` (where scope prefixed constants and constants not found in the
local scope are sought) Much code had to change in order to correctly
thread these two through the system and ensure that they were correctly
set.  Most significantly the `responderObject` function needs to be passed
these as arguments.  A change that rippled widely throughout the code.
One significant needed alteration was to place `spawn` into the EvalM
monad so it could access the context.  A special case, called
`spawnObject` was retained in the IO monad so that we could boot the
system.  It assumes that Object will be in its own scope and be its own
global.

An addition to the graph search algorithm has been made to look for
constants (defined as beginning with a unicode capital letter) in the new
scopes.  It also required adding a `setIVar` function to class.

Objects can now be compared (at a very low level) by means of a TMVar that
holds their process ID.  We note that this could allow Spawn.hs to
directly set the process field, but we have left in in the initialization
function of the spawned object so as to make hooks easier to implement.