Added Error Class and Initialize WIP
We have added the Error class as base class for error results.

We have added a call to the `initialize` method when objects are created.
Unfortunatly, the version for `new` is having trouble finding the method.  This
is likely related to a simular bug that is preventing testfile4 from running
because of a missing `+` method.

Next step is to rig up object promotion to create an Error instance.

(Personal Note: this commit made from St. Dorothy's Rest in Camp Meeker
California.)