Bug Fix: Arguments Evaluated in Wrong Context
The arguments to method calls to local variables were being evaluated in the
context of the reciever not the context where they were created.  This was
causing a number of odd `nil`s to show up in a number of places.  Because
OpStrings get reduced to method calles this was also showing up in any equation
that involved a non-primitive argument for its secong operand.