Special Form `super` Implimented (Closes #48)
The new keyword super is ised t invoke the method which would have been
next invoked had the current method not been prenent.  This can be used by
subclasses to do some work while also invoking the original method.

Potential issues:
Because of the way super is implimented, any state in the active thread is
not subject to change and will be invoked as it existed when the current
method started.  In poarticular this means tha new modules added to self
withing a method are not subjet to search when super is invoked.