[WIP] Parameters and Scope
Eval.Parameters for matching arguments with formal parameters.

Scope to be used by eval to abstract the function of the underlying monad
as, for example, is done in Test/Parameters.hs.

Name.hs is repository for common type definations not dependent on any other
data structures.  Will need to integrate this into other files as they are
are updated.

Updated some copyright notices.