Allow for mutation of Runtime Function
Previously the assumption was that a Process would keep the same evaluation
function for its entire lifetime.  It turns out that there are some edge cases
where it is convient to actually alter that function. This change allows for
that to happen by using the resulting RTS field rather then looping forever on
the initial function passed.  So to change functions one must simply update
the RTS state while running in the Runtime monad.  (This requires access to
the "deep" state which means it cannot be done in a Scope moand opperation
like eval"

The primary place this is wanted it in bootstrapping.  Setting up the initial
loop of Object and Class requires one to be told of the other only after
being formed.  Once done both of these Classes should behave as any other
class.

It is also useful for forming a class which has an initial action in the form
of an eval block. The behavior of the proceess while evaluationg that block is
different than the behavior once it is setup.