newScope instance & Addational Framework for Boot
Still pluggig away at the Scope instance for Runtime.  This time with the
newScope function.  Care should be take as the this function will break tail
recursion.  It may be necesary to expose (and probably rename) the internal
function `ap` as it does not break tail calls.

Boot is not yet done, but the scaffolding is there.  We create a new process
and send in the "initialize" method call.  Some of this may want to be
refactored back into Runtime.

Still to be done is writing the actuall boot up code, but at least now there
is a place to put it.