Function Signature Changed to Response
Functions no longer return an `Object` but a `Response`.

A response is a dummy data type that simply asserts that the medhod has in
some way provided for responding to the caller either by replying directly
or by properly calling a different method using tailCall (or the specialized
tail call `super`).

Withing a Scope the only way to find a Response value is to call one of
these functions, and becase we require that a function return this value, we
can statically type check that someone will respond. This should cut down on
the number of "Blocked Indefinatly on ..." errors.

See [this StackOverflow question][1].

[1]: http://stackoverflow.com/questions/43382213/a-monadtransformer-to-typecheck-for-a-completed-action