Function `send` now actually works
Replaced the defination of send with something that did not allow for
arbitrary action in the Runtime monad, but simply provided the option to
send a message with a name of the users choosing and the value of the
resulting computation to a designated process.

This premitted the use of low level operations and solve the problem of
having to produce the state in which the arbitrary actiosn should take
place.